api/requirements.txt
fastapi
uvicorn
promptflow-core
python-dotenv